% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{calculate_sedc}
\alias{calculate_sedc}
\title{Calculate SEDC covariates}
\usage{
calculate_sedc(
  point_from,
  point_to,
  id,
  sedc_bandwidth,
  threshold,
  target_fields
)
}
\arguments{
\item{point_from}{SpatVector object. Locations where the sum of SEDCs are calculated.}

\item{point_to}{SpatVector object. Locations where each SEDC is calculated.}

\item{id}{character(1). Name of the unique id field in point_to.}

\item{sedc_bandwidth}{numeric(1).
Distance at which the source concentration is reduced to
exp(-3) (approximately -95 \%)}

\item{threshold}{numeric(1). For computational efficiency,
the nearest points in threshold will be selected.
Default is \code{2 * sedc_bandwidth}.}

\item{target_fields}{character(varying). Field names in characters.}
}
\description{
Calculate SEDC covariates
}
\note{
sf implementation is pending. Only available for terra.
Currently the function internally convert sf objects to terra.
}
\author{
Insang Song
}