% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_crs2}
\alias{check_crs2}
\title{check_crs2: Coordinate system checker}
\usage{
check_crs2(input, crs_standard = "EPSG:4326")
}
\arguments{
\item{input}{Input object one of sf or terra::Spat* object}

\item{crs_standard}{character(1). A standard definition of coordinate reference system. Default is "EPSG:4326" Consult \href{https://epsg.io}{epsg.io} for details of other CRS.}
}
\value{
A (reprojected) sf or SpatVector object.
}
\description{
The input is checked whether its coordinate system is present. If not, it is reprojected to EPSG:5179.
}
