% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_crs_align}
\alias{check_crs_align}
\title{check_crs_align: Check coordinate system then reproject}
\usage{
check_crs_align(input, crs_standard = "EPSG:4326")
}
\arguments{
\item{input}{Input object one of sf or terra::Spat* object}

\item{crs_standard}{character(1). A standard definition of
coordinate reference system. Default is "EPSG:4326"
Consult \href{https://epsg.io}{epsg.io} for details of other CRS.}
}
\value{
A (reprojected) sf or SpatVector object.
}
\description{
The input is checked whether its coordinate system is
present. If not, it is reprojected to the CRS specified in
\code{crs_standard}.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

base_crs <- "OGC:CRS84"
nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
check_crs_align(nc_sf, base_crs)

nc_vect <- terra::vect(nc_sf)
check_crs_align(nc_vect, base_crs)
}
\author{
Insang Song
}