% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_datatype}
\alias{check_datatype}
\title{Return the data type}
\usage{
check_datatype(input)
}
\arguments{
\item{input}{Spat*/sf/stars object.}
}
\value{
character(1). One of 'vector' or 'raster'.
}
\description{
This function returns one of 'vector' or 'raster'
depending on the input class.
}
\note{
Although \code{stars} object is a little ambiguous
whether to classify vector or raster,
it will be considered raster in this package.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
check_packbound(nc_sf)
nc_vect <- terra::vect(nc_sf)
check_packbound(nc_vect)
}
\author{
Insang Song
}