% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_packbound}
\alias{check_packbound}
\title{Return the package the input object is based on}
\usage{
check_packbound(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
A character object; one of 'terra' and 'sf'
}
\description{
Detect whether the input object is sf or Spat* object.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
check_packbound(nc_sf)
nc_vect <- terra::vect(nc_sf)
check_packbound(nc_vect)
## END OF EXAMPLE
}
\author{
Insang Song
}