% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_subject}
\alias{check_subject}
\title{Check the subject object and perform necessary conversions if needed.}
\usage{
check_subject(subject, extent = NULL, subject_id = NULL)
}
\arguments{
\item{subject}{The input object to be checked.}

\item{extent}{The extent of the subject object.}

\item{subject_id}{Optional. The ID of the subject object.}
}
\value{
The checked and converted subject object.
}
\description{
This function checks the class of the input object and
performs necessary conversions if needed.
}
\examples{
# Check a SpatVector object
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
ncsf <- sf::st_read(ncpath)
extent <- c(-80, -77, 35, 36)
check_subject(subject = ncsf, extent = extent, subject_id = "FIPS")

# Check a character object
check_subject(subject = ncpath, extent = extent, subject_id = "FIPS")
}
\keyword{internal}
