% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopinOptions.R
\name{chopinOptions}
\alias{chopinOptions}
\title{Internal function for setting chopin options}
\usage{
chopinOptions(chopin.backend = "mirai")
}
\arguments{
\item{chopin.backend}{The backend to be set. Default is "mirai".}
}
\description{
Internal function for setting chopin options
}
\keyword{internal}
