% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{clip_as_extent_ras}
\alias{clip_as_extent_ras}
\title{clip_as_extent_ras: Clip input raster.}
\usage{
clip_as_extent_ras(pnts = NULL, buffer_r = NULL, ras = NULL, nqsegs = 180L)
}
\arguments{
\item{pnts}{\code{sf} or \code{SpatVector} object}

\item{buffer_r}{numeric(1). buffer radius.
This value will be automatically multiplied by 1.25}

\item{ras}{\code{SpatRaster} object to be clipped}

\item{nqsegs}{\code{integer(1)}. the number of points per a quarter circle}
}
\description{
Clip input raster by the expected maximum extent of
computation.
}
\examples{
library(terra)

ras_rand <- terra::rast(nrow = 20, ncol = 20)
terra::values(ras_rand) <- runif(400L)
ras_rand_p <-
  data.frame(
    x = c(3, 5, 3.2, 8),
    y = c(12, 10, 15, 12),
    z = c(0, 1, 2, 3)
  )
ras_rand_p <- terra::vect(ras_rand_p, geom = c("x", "y"))
clip_as_extent_ras(ras_rand_p, 1.5, ras_rand)
}
\author{
Insang Song
}