% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{clip_ras_ext}
\alias{clip_ras_ext}
\title{Clip input raster with a buffered vector extent.}
\usage{
clip_ras_ext(x = NULL, y = NULL, radius = NULL, nqsegs = 180L, extrusion = 1.1)
}
\arguments{
\item{x}{\code{SpatRaster} object to be clipped}

\item{y}{\code{sf} or \code{SpatVector} object}

\item{radius}{numeric(1). buffer radius.
This value will be automatically multiplied by \code{extrusion}}

\item{nqsegs}{integer(1). the number of points per a quarter circle.
Default is 180L.}

\item{extrusion}{numeric(1). Extrusion factor for the extent. Default is 1.1.}
}
\value{
A clipped \code{SpatRaster} object.
}
\description{
Clip input raster by the expected maximum extent of
computation.
}
\seealso{
Other Helper functions: 
\code{\link{clip_vec_ext}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{par_def_q}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
\keyword{soft-deprecated}
