% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{clip_vec_ext}
\alias{clip_vec_ext}
\title{Clip to the buffered extent of input vector}
\usage{
clip_vec_ext(pnts, radius, target_input)
}
\arguments{
\item{pnts}{\code{sf} or \code{SpatVector} object}

\item{radius}{\code{numeric(1)}. Circular buffer radius.
this value will be automatically multiplied by 1.1}

\item{target_input}{\code{sf} or \code{SpatVector} object to be clipped}
}
\value{
A clipped \code{sf} or \code{SpatVector} object.
}
\description{
Clip input vector by
the expected maximum extent of computation.
}
\examples{
library(sf)
library(stars)
library(terra)
options(sf_use_s2 = FALSE)

bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
bcsd <- stars::read_stars(bcsd_path)
bcsd_rpnt <- sf::st_sample(bcsd, 4L)
bcsd_rpntm <- sf::st_sample(bcsd, 1000L)
clip_vec_ext(bcsd_rpntm, 1000, bcsd_rpnt)
}
\author{
Insang Song
}