% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{cut_coords}
\alias{cut_coords}
\title{Partition coordinates into quantiles}
\usage{
cut_coords(x = NULL, y = NULL, quantiles)
}
\arguments{
\item{x}{numeric/sf/SpatVector. x-coordinates (if numeric).}

\item{y}{numeric. y-coordinates.}

\item{quantiles}{numeric vector. Quantiles.}
}
\value{
A data.frame of x-y quantile pairs.
}
\description{
Partition coordinates into quantiles
}
\note{
This function is only for two-dimensional points.
}
\examples{
random_points <-
  data.frame(x = runif(1000, 0, 100), y = runif(1000, 0, 100))
quantiles <- qdef(4L)
cut_coords(random_points$x, random_points$y, quantiles)
}
\keyword{internal}