% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{datamod}
\alias{datamod}
\title{Return the input's GIS data model type}
\usage{
datamod(input)
}
\arguments{
\item{input}{Spat*/sf/stars object.}
}
\value{
character(1). One of \code{"vector"} or \code{"raster"}.
}
\description{
This function returns one of 'vector' or 'raster'
depending on the input class.
}
\note{
Although \code{stars} object is a little ambiguous
whether to classify vector or raster,
it will be considered raster in this package.
}
\seealso{
Other Helper functions:
\code{\link[=dep_check]{dep_check()}},
\code{\link[=dep_switch]{dep_switch()}},
\code{\link[=get_clip_ext]{get_clip_ext()}},
\code{\link[=par_def_q]{par_def_q()}},
\code{\link[=reproject_std]{reproject_std()}},
\code{\link[=reproject_to_raster]{reproject_to_raster()}}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
