% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dep_switch}
\alias{dep_switch}
\title{Switch spatial data class}
\usage{
dep_switch(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
Data converted to the other package class
(if sf, terra; if terra, sf)
}
\description{
Convert class between \code{sf}/\code{stars}-\code{terra}
}
\seealso{
Other Helper functions:
\code{\link[=datamod]{datamod()}},
\code{\link[=dep_check]{dep_check()}},
\code{\link[=get_clip_ext]{get_clip_ext()}},
\code{\link[=par_def_q]{par_def_q()}},
\code{\link[=reproject_std]{reproject_std()}},
\code{\link[=reproject_to_raster]{reproject_to_raster()}}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
