% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{detect_class}
\alias{detect_class}
\title{Detect classes in function arguments}
\usage{
detect_class(args = NULL, search = NULL)
}
\arguments{
\item{args}{Any list, but preferably generated by \code{list(...)} inside
a function.}

\item{search}{character(1). Class name to search. Partial match is supported.}
}
\value{
logical vector.
}
\description{
When a R function is defined in an ordinary
fashion (i.e., assigning a function by \code{<- function(...)})
would be subject to ambiguity particularly if the function
name is the same as the generic function name(s).
This function supports detecting classes of arguments in
a loosely defined function.
}
\examples{
df <- data.frame(a = 1, b = 3)
detect_class(list(df), "data.frame")
}
\author{
Insang Song
}