% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_backend.R
\name{.backend_collector}
\alias{.backend_collector}
\title{Internal function for collecting parallel worker results}
\usage{
.backend_collector(worker)
}
\arguments{
\item{worker}{Unresolved parallel worker}
}
\description{
Internal function for collecting parallel worker results
}
\keyword{internal}
