% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_backend.R
\name{.backend_worker}
\alias{.backend_worker}
\title{backend_worker function}
\usage{
.backend_worker(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the backend function.}
}
\value{
The result of the backend function call.
}
\description{
This function is used to execute a specific backend based on the value of the "chopin.backend" option.
If the backend is set to "mirai", the function calls the mirai::mirai() function.
If the backend is set to "future", the function calls the future::future() function with seed and lazy options enabled.
If the backend is set to any other value, the function throws an error indicating an unknown backend.
}
\keyword{internal}
