% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.check_character}
\alias{.check_character}
\title{Check the class of an input object}
\usage{
.check_character(input)
}
\arguments{
\item{input}{The input object to be checked}
}
\value{
A character string indicating the class of the input object ("raster" or "vector")
}
\description{
This function checks the class of an input object and returns "raster" if it is a raster object,
or "vector" if it is a vector object.
}
\keyword{internal}
