% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.check_raster}
\alias{.check_raster}
\title{Check Raster Input}
\usage{
.check_raster(input, extent = NULL)
}
\arguments{
\item{input}{The input object to be checked. It can be either
a SpatRaster object or a character path to a raster file.}

\item{extent}{The extent of the raster. Defaults to NULL.
Numeric vector should be put in order of
\code{c(xmin, xmax, ymin, ymax)}.}
}
\value{
The validated input object.
}
\description{
This function checks the input object to ensure
it is a valid raster object or a character path to a raster file.
It also provides warnings and informative messages based on the input type.
}
\examples{
.check_raster(system.file("extdata/nc_srtm15_otm.tif", package = "chopin"))

}
\keyword{internal}
