% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.check_vector}
\alias{.check_vector}
\title{Check the subject object and perform necessary conversions if needed.}
\usage{
.check_vector(
  input,
  input_id = NULL,
  extent = NULL,
  out_class = character(1),
  ...
)
}
\arguments{
\item{input}{sf/SpatVector/character. The input object to be checked.}

\item{input_id}{character(1). ID field of the subject object.}

\item{extent}{numeric(4). The extent of the subject object.
Numeric vector should be put in order of
\code{c(xmin, xmax, ymin, ymax)}.}

\item{out_class}{character(1). The class of the output object.
Should be one of \code{c("sf", "terra")}.}

\item{...}{Placeholder.}
}
\value{
The checked and converted subject object.
}
\description{
This function checks the class of the input object and
performs necessary conversions if needed.
}
\keyword{internal}
