% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.intersect_extent}
\alias{.intersect_extent}
\alias{.intersect_extent,sf,ANY-method}
\alias{.intersect_extent,SpatExtent,ANY-method}
\alias{.intersect_extent,SpatVector,ANY-method}
\alias{.intersect_extent,numeric,character-method}
\title{Get intersection extent}
\usage{
.intersect_extent(input, out_class, ...)

\S4method{.intersect_extent}{sf,ANY}(input, out_class = NULL, ...)

\S4method{.intersect_extent}{SpatExtent,ANY}(input, out_class = NULL, ...)

\S4method{.intersect_extent}{SpatVector,ANY}(input, out_class = NULL, ...)

\S4method{.intersect_extent}{numeric,character}(input, out_class = NULL, ...)
}
\arguments{
\item{input}{sf/SpatExtent/SpatVector/numeric}

\item{out_class}{character(1). "sf" or "terra"}

\item{...}{other arguments. Placeholder.}
}
\description{
Get intersection extent
}
\keyword{internal}
