% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R, R/scale_process_mirai.R
\name{.par_screen}
\alias{.par_screen}
\title{Prescreen input data for parallelization}
\usage{
.par_screen(type, input, input_id = NULL, out_class = "terra", .window = NULL)

.par_screen(type, input, input_id = NULL, out_class = "terra", .window = NULL)
}
\arguments{
\item{type}{character(1). "raster" or "vector".}

\item{input}{object. Input object.}

\item{input_id}{character(1). Default is NULL. If NULL, the function
will not check the object with an ID column.}

\item{out_class}{character(1). Default is NULL, but should be one of
\code{c("sf", "terra")}. Default is "terra".}

\item{.window}{numeric(4)/SpatExtent/st_bbox object. Loading window.}
}
\description{
This function takes input object and type character to ingest
the input object to return the object in the desired class.

This function takes input object and type character to ingest
the input object to return the object in the desired class.
}
\keyword{internal}
