% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{estimate_demands}
\alias{estimate_demands}
\title{Estimate computational demands from inputs (to be written)}
\usage{
estimate_demands(inputs, nx, ny, padding)
}
\arguments{
\item{inputs}{character vector of file paths}

\item{nx}{integer(1).}

\item{ny}{integer(1).}

\item{padding}{numeric(1). Extrusion factor}
}
\description{
Estimate computational demands from inputs (to be written)
}
\author{
Insang Song
}