% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{ext_to_poly}
\alias{ext_to_poly}
\title{Generate a rectangular polygon from extent}
\usage{
ext_to_poly(extent = NULL, output_class = c("sf", "terra"), crs = "EPSG:4326")
}
\arguments{
\item{extent}{input extent.
A numeric vector with xmin/xmax/ymin/ymax,
\link[sf:st_bbox]{sf::st_bbox} or \link[terra:ext]{terra::ext} outputs.}

\item{output_class}{character(1).
Class of the output polygon. One of \code{"sf"} or \code{"terra"}}

\item{crs}{character(1). Coordinate reference system definition.}
}
\value{
\code{sf} or \code{SpatVector} object of a rectangular polygon.
}
\description{
Generate a rectangular polygon from extent
}
\examples{
library(sf)
library(terra)
numext1 <- c(-100, -70, 30, 40)
names(numext1) <- c("xmin", "xmax", "ymin", "ymax")
ext_to_poly(numext1, "sf")
ext_to_poly(numext1, "terra")
}
\seealso{
Other Helper functions: 
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{par_def_q}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
\keyword{soft-deprecated}
