% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extent_to_polygon}
\alias{extent_to_polygon}
\title{Generate a rectangular polygon from extent}
\usage{
extent_to_polygon(
  extent = NULL,
  output_class = c("sf", "terra"),
  crs = "EPSG:4326"
)
}
\arguments{
\item{extent}{input extent.
A numeric vector with xmin/xmax/ymin/ymax,
\link[sf:st_bbox]{sf::st_bbox} or \link[terra:ext]{terra::ext} outputs.}

\item{output_class}{character(1).
Class of the output polygon. One of \code{"sf"} or \code{"terra"}}

\item{crs}{character(1). Coordinate reference system definition.}
}
\value{
\code{sf} or \code{SpatVector} object of a rectangular polygon.
}
\description{
Generate a rectangular polygon from extent
}
\examples{
library(sf)
library(terra)
numext1 <- c(-100, -70, 30, 40)
names(numext1) <- c("xmin", "xmax", "ymin", "ymax")
extent_to_polygon(numext1, "sf")
extent_to_polygon(numext1, "terra")
}
\author{
Insang Song
}