% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_at_buffer}
\alias{extract_at_buffer}
\alias{extract_at_buffer_flat}
\alias{extract_at_buffer_kernel}
\title{Extract summarized values from raster with points and a buffer radius}
\usage{
extract_at_buffer(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = 90L,
  func = "mean",
  kernel = NULL,
  bandwidth = NULL,
  extent = NULL,
  max_cells = 2e+07,
  ...
)

extract_at_buffer_flat(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = NULL,
  func = "mean",
  kernel = NULL,
  bandwidth = NULL,
  max_cells = 2e+07,
  ...
)

extract_at_buffer_kernel(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = NULL,
  func = stats::weighted.mean,
  kernel = NULL,
  bandwidth = NULL,
  max_cells = 2e+07,
  ...
)
}
\arguments{
\item{points}{\code{sf}/\code{SpatVector} object.
Coordinates where buffers will be generated.}

\item{surf}{\code{SpatRaster} object or file path(s) with extensions
that are GDAL-compatible. A raster from which a summary will be calculated}

\item{radius}{numeric(1). Buffer radius.
Here we assume circular buffers only}

\item{id}{character(1). Unique identifier of each point.}

\item{qsegs}{integer(1). Number of vertices at a quarter of a circle.
Default is \code{90L}.}

\item{func}{a function taking a numeric vector argument.}

\item{kernel}{character(1). Name of a kernel function
One of \code{"uniform"}, \code{"triweight"}, \code{"quartic"}, and \code{"epanechnikov"}}

\item{bandwidth}{numeric(1). Kernel bandwidth.}

\item{extent}{numeric(4) or SpatExtent. Extent of clipping vector.
It only works with \code{points} of character(1) file path.
When using numeric(4), it should be in the order of
\code{c(xmin, xmax, ymin, ymax)}. The coordinate system should be the same
as the \code{points}.}

\item{max_cells}{integer(1). Maximum number of cells in memory.
See \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract}} for more details.}

\item{...}{Placeholder.}
}
\value{
a data.frame object with mean value
}
\description{
For simplicity, it is assumed that the coordinate systems of
the points and the raster are the same.
}
\note{
When \code{Sys.setenv("CHOPIN_FORCE_CROP" = "TRUE")} is set, the raster will be
cropped to the extent of the polygons (with \code{snap} = \code{"out"}).
To note, the function is designed to work with the \code{exactextractr} package.
Arguments of \code{exactextractr::exact_extract} are set as below
(default otherwise listed):
\itemize{
\item \code{force_df} = \code{TRUE}
\item \code{stack_apply} = \code{TRUE}
\item \code{max_cells_in_memory} = \code{2e8}
\item \code{progress} = \code{FALSE}
}
}
\examples{
library(terra)
rrast <- terra::rast(nrow = 100, ncol = 100)
terra::crs(rrast) <- "EPSG:5070"
terra::values(rrast) <- rgamma(1e4, 4, 2)
rpnt <- terra::spatSample(rrast, 100L, as.points = TRUE)
rpnt$pid <- sprintf("id_\%03d", seq(1, 100))
extract_at_buffer(rpnt, rrast, 4, "pid")
}
\seealso{
Other Macros for calculation: 
\code{\link{extract_at}()},
\code{\link{extract_at_poly}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_aw}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}
