% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_at_poly}
\alias{extract_at_poly}
\title{Extract summarized values from raster with generic polygons}
\usage{
extract_at_poly(
  polys = NULL,
  surf = NULL,
  id = NULL,
  func = "mean",
  max_cells = 2e+07,
  ...
)
}
\arguments{
\item{polys}{\code{sf}/\code{SpatVector} object. Polygons.}

\item{surf}{\code{SpatRaster} object or file path(s) with extensions
that are GDAL-compatible. A raster from which a summary will be calculated}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{a generic function name in string or
a function taking two arguments that are
compatible with \code{\link[exactextractr]{exact_extract}}.
For example, \code{"mean"} or \verb{\\(x, w) weighted.mean(x, w, na.rm = TRUE)}}

\item{max_cells}{integer(1). Maximum number of cells in memory.
See \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract}} for more details.}

\item{...}{Placeholder.}
}
\value{
a data.frame object with function value
}
\description{
For simplicity, it is assumed that the coordinate systems of
the points and the raster are the same.
}
\note{
When \code{Sys.setenv("CHOPIN_FORCE_CROP" = "TRUE")} is set, the raster will be
cropped to the extent of the polygons (with \code{snap} = \code{"out"}).
To note, the function is designed to work with the \code{exactextractr} package.
Arguments of \code{exactextractr::exact_extract} are set as below
(default otherwise listed except for max_cells_in_memory,
which is set in the \code{max_cells} argument):
\itemize{
\item \code{force_df} = \code{TRUE}
\item \code{stack_apply} = \code{TRUE}
\item \code{progress} = \code{FALSE}
}
}
\examples{
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
nc <- terra::project(nc, "EPSG:5070")
rrast <- terra::rast(nc, nrow = 100, ncol = 220)
ncr <- terra::rasterize(nc, rrast)
terra::values(rrast) <- rgamma(2.2e4, 4, 2)
rpnt <- terra::spatSample(rrast, 16L, as.points = TRUE)
rpnt$pid <- sprintf("ID-\%02d", seq(1, 16))
rpoly <-
  terra::buffer(rpnt, 5, capstyle = "square", joinstyle = "bevel")
extract_at_poly(rpoly, rrast, "pid")
}
\seealso{
Other Macros for calculation: 
\code{\link{extract_at}()},
\code{\link{extract_at_buffer}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_aw}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}
