% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extract_with}
\alias{extract_with}
\title{Extract raster values with point buffers or polygons}
\usage{
extract_with(
  vector = NULL,
  raster = NULL,
  id = NULL,
  func = "mean",
  mode = c("polygon", "buffer"),
  ...
)
}
\arguments{
\item{vector}{\code{sf}/\code{SpatVector} object.}

\item{raster}{\code{SpatRaster} object.}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{function taking one numeric vector argument.}

\item{mode}{one of \code{"polygon"}
(generic polygons to extract raster values with) or
\code{"buffer"} (point with buffer radius)}

\item{...}{various. Passed to extract_with_buffer.
See \code{?extract_with_buffer} for details.}
}
\value{
A data.frame object with summarized raster values with
respect to the mode (polygon or buffer) and the function.
}
\description{
Extract raster values with point buffers or polygons
}
\examples{
## See ?extract_with_polygons and ?extract_with_buffers
}
\author{
Insang Song \email{geoissong@gmail.com}
}