% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{get_computational_regions}
\alias{get_computational_regions}
\title{Get a set of computational regions}
\usage{
get_computational_regions(
  input,
  mode = c("grid", "grid_advanced", "density"),
  nx = 10L,
  ny = 10L,
  grid_min_features = 30L,
  padding = NULL,
  unit = NULL,
  ...
)
}
\arguments{
\item{input}{sf or Spat* object.}

\item{mode}{character(1). Mode of region construction. One of "grid" (simple grid regardless of the number of features in each grid), "density" (clustering-based varying grids), "grid_advanced" (merging adjacent grids with smaller number of features than grid_min_features).}

\item{nx}{integer(1). The number of grids along x-axis.}

\item{ny}{integer(1). The number of grids along y-axis.}

\item{grid_min_features}{integer(1). A threshold to merging adjacent grids}

\item{padding}{numeric(1). A extrusion factor to make buffer to
clip actual datasets. Depending on the length unit of the CRS of input.}

\item{unit}{character(1). The length unit for padding (optional).
units::set_units is used for padding when sf object is used.
See \href{https://cran.r-project.org/web/packages/units/vignettes/measurement_units_in_R.html}{units package vignette (web)}
for the list of acceptable unit forms.}

\item{...}{arguments passed to the internal function}
}
\value{
A list of two,
\code{original}: exhaustive and non-overlapping grid polygons in the class of input
\code{padded}: a square buffer of each polygon in \code{original}. Used for computation.
}
\description{
Using input points, the bounding box is split to
the predefined numbers of columns and rows. Each grid will be buffered by the radius.
}
\examples{
# data
library(sf)
ncpath <- system.file("shape/nc.shp", package = "sf")
nc <- read_sf(ncpath)
nc <- st_transform(nc, "EPSG:5070")
# run
# nc_comp_region <- get_computational_regions(nc, nx = 12, ny = 8)

}
\author{
Insang Song
}