% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{methods}
\name{indexing}
\alias{indexing}
\alias{`[`}
\alias{.intersect}
\title{Subset for nonidentical package class objects}
\usage{
\S4method{[}{SpatVector,bbox,missing,ANY}(x, i, j)

\S4method{[}{SpatVector,sf,missing,ANY}(x, i, j)

\S4method{[}{SpatVector,sfc,missing,ANY}(x, i, j)

\S4method{[}{SpatVector,SpatExtent,missing,ANY}(x, i, j)

.intersect(x, y)
}
\arguments{
\item{x}{SpatVector/sf/SpatRaster object to be intersected.}

\item{i}{Dataset used to subset x.}

\item{j}{Column indices or names.}

\item{y}{SpatVector/sf object. Intersecting object.}
}
\description{
Subset for nonidentical package class objects

Intersect different data model objects
}
\keyword{internal}
