% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{is_bbox_within_reference}
\alias{is_bbox_within_reference}
\title{Check if the data extent is inside the reference bounding box}
\usage{
is_bbox_within_reference(data_query = NULL, reference = NULL)
}
\arguments{
\item{data_query}{sf*/stars/SpatVector/SpatRaster object.}

\item{reference}{sf*/stars/SpatVector/SpatRaster object}
}
\value{
logical(1). \code{TRUE} (the queried data extent is completely within
the reference bounding box) or \code{FALSE}
}
\description{
One of the most common errors in spatial computation is rooted
in the entirely or partly incomparable spatial extents of input datasets.
This function returns whether your data is inside the target computational
extent.
It is assumed that you know and have the exact computational region.
This function will return \code{TRUE} if the reference region
completely contains your data's extent and \code{FALSE} otherwise.
}
\examples{
library(sf)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- sf::st_read(ncpath)
nc <- sf::st_transform(nc, "EPSG:4326")

refextnum <- c(-100, -60, 20, 40)
names(refextnum) <- c("xmin", "xmax", "ymin", "ymax")
refext <- ext2poly(refextnum)
is_bbox_within_reference(nc, refext)
}
\author{
Insang Song \email{geoissong@gmail.com}
}