% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{is_within_ref}
\alias{is_within_ref}
\title{Check if the boundary of the vector/raster object is inside the reference}
\usage{
is_within_ref(input_object, reference)
}
\arguments{
\item{input_object}{sf/stars/SpatVector/SpatRaster object.}

\item{reference}{sf/stars/SpatVector/SpatRaster object.}
}
\value{
logical
}
\description{
Check if the boundary of the vector/raster object is inside the reference
}
\examples{
library(sf)
sf_use_s2(FALSE)
ncpath <- system.file("shape/nc.shp", package = "sf")
nc <- sf::read_sf(ncpath)
nc <- sf::st_transform(nc, "EPSG:4326")
mainland_vec <- c(xmin = -128, xmax = -62, ymin = 22, ymax = 52)
mainland_box <- ext2poly(mainland_vec, output_class = "sf")
within_res <- is_within_ref(nc, mainland_box)
within_res
}
\author{
Insang Song \email{geoissong@gmail.com}
}