% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowrap.R
\name{par_compose}
\alias{par_compose}
\title{Auto-parallelize}
\usage{
par_compose(fun, x, y, strategy, ..., .backend = c("future", "mirai"))
}
\arguments{
\item{fun}{function.}

\item{x}{X argument of \code{fun}}

\item{y}{Y argument of \code{fun}}

\item{strategy}{character(1). Parallel processing strategy
supported in \code{chopin}}

\item{...}{Other arguments to be passed to \code{fun}}

\item{.backend}{character(1). One of "future" or "mirai"}
}
\value{
data.frame
}
\description{
Auto-parallelize
}
