% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridding.R
\name{par_cut_coords}
\alias{par_cut_coords}
\title{Partition coordinates into quantile polygons}
\usage{
par_cut_coords(x = NULL, y = NULL, quantiles)
}
\arguments{
\item{x}{numeric/sf/SpatVector. x-coordinates (if numeric).}

\item{y}{numeric. y-coordinates.}

\item{quantiles}{numeric vector. Quantiles.}
}
\value{
A \code{SpatVector} object with field \code{CGRIDID}.
}
\description{
Partition coordinates into quantile polygons
}
\note{
This function is only for two-dimensional points.
}
\seealso{
Other Parallelization:
\code{\link[=par_grid]{par_grid()}},
\code{\link[=par_grid_mirai]{par_grid_mirai()}},
\code{\link[=par_hierarchy]{par_hierarchy()}},
\code{\link[=par_hierarchy_mirai]{par_hierarchy_mirai()}},
\code{\link[=par_make_dggrid]{par_make_dggrid()}},
\code{\link[=par_make_grid]{par_make_grid()}},
\code{\link[=par_make_h3]{par_make_h3()}},
\code{\link[=par_merge_grid]{par_merge_grid()}},
\code{\link[=par_multirasters]{par_multirasters()}},
\code{\link[=par_multirasters_mirai]{par_multirasters_mirai()}},
\code{\link[=par_pad_balanced]{par_pad_balanced()}},
\code{\link[=par_pad_grid]{par_pad_grid()}},
\code{\link[=par_split_list]{par_split_list()}}
}
\concept{Parallelization}
\keyword{internal}
