% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridding.R
\name{par_def_q}
\alias{par_def_q}
\title{Quantile definition}
\usage{
par_def_q(steps = 4L)
}
\arguments{
\item{steps}{integer(1). The number of quantiles.}
}
\value{
numeric vector of quantiles.
}
\description{
Quantile definition
}
\seealso{
Other Helper functions:
\code{\link[=datamod]{datamod()}},
\code{\link[=dep_check]{dep_check()}},
\code{\link[=dep_switch]{dep_switch()}},
\code{\link[=get_clip_ext]{get_clip_ext()}},
\code{\link[=reproject_std]{reproject_std()}},
\code{\link[=reproject_to_raster]{reproject_to_raster()}}
}
\concept{Helper functions}
\keyword{internal}
