% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R
\name{par_fallback}
\alias{par_fallback}
\title{Parallelization error fallback}
\usage{
par_fallback(err = NULL, fun = NULL, inputid = NULL)
}
\arguments{
\item{err}{Error status or message.}

\item{fun}{function.}

\item{inputid}{character(1). ID of the computational region.
For example, \code{par_make_gridset} output should have \code{"CGRIDID"},
which will be included in the outcome data.frame.
If the function does not have an argument containing \code{"id"},
the output will have a column named \code{"chopin_domain_id"}.}
}
\value{
data.frame with domain id and error message.
}
\description{
Parallelization error fallback
}
\note{
This function assumes that the \code{fun} has an argument containing
\code{"id"}.
}
\examples{
err <- simpleError("No input.")
par_fallback(err, extract_at, inputid = 1)
}
\seealso{
Other Parallelization: 
\code{\link{par_cut_coords}()},
\code{\link{par_grid}()},
\code{\link{par_group_grid}()},
\code{\link{par_hierarchy}()},
\code{\link{par_make_grid}()},
\code{\link{par_make_gridset}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()}
}
\author{
Insang Song
}
\concept{Parallelization}
