% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{qdef}
\alias{qdef}
\title{Quantile definition}
\usage{
qdef(steps = 4L)
}
\arguments{
\item{steps}{integer(1). The number of quantiles.}
}
\value{
numeric vector of quantiles.
}
\description{
Quantile definition
}
\keyword{internal}