% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{rast_short}
\alias{rast_short}
\title{Quick call for SpatRaster with a window}
\usage{
rast_short(rasterpath = NULL, win = NULL)
}
\arguments{
\item{rasterpath}{character(1). Path to the raster file.}

\item{win}{Named integer vector (4) or terra::ext() results.}
}
\value{
SpatRaster object.
}
\description{
Quick call for SpatRaster with a window
}
\examples{
library(terra)
bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
ext_small <- terra::ext(
  c(xmin = -80, xmax = -76, ymin = 35, ymax = 36)
)
rast_short(bcsd_path, ext_small)
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext2poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}