% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{reproject_b2r}
\alias{reproject_b2r}
\title{Reproject vectors to raster's CRS}
\usage{
reproject_b2r(vector, raster)
}
\arguments{
\item{vector}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object}

\item{raster}{\code{SpatRaster} object}
}
\value{
Reprojected object in the same class as \code{vector}
}
\description{
Reproject vectors to raster's CRS
}
\examples{
# NOT TO RUN
}
\author{
Insang Song
}