% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{reproject_to_raster}
\alias{reproject_to_raster}
\title{Align vector CRS to raster's}
\usage{
reproject_to_raster(vector = NULL, raster = NULL)
}
\arguments{
\item{vector}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object}

\item{raster}{\code{SpatRaster} object}
}
\value{
Reprojected object in the same class as \code{vector}
}
\description{
Align vector CRS to raster's
}
\seealso{
Other Helper functions:
\code{\link[=datamod]{datamod()}},
\code{\link[=dep_check]{dep_check()}},
\code{\link[=dep_switch]{dep_switch()}},
\code{\link[=get_clip_ext]{get_clip_ext()}},
\code{\link[=par_def_q]{par_def_q()}},
\code{\link[=reproject_std]{reproject_std()}}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
