% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\docType{package}
\name{scomps-package}
\alias{scomps}
\alias{scomps-package}
\title{scomps: Scalable R geospatial computation}
\description{
A package for scalable geospatial computation for environmental health research
}
\author{
\strong{Maintainer}: Insang Song \email{geoissong@gmail.com} (\href{https://orcid.org/0000-0001-8732-3256}{ORCID})

}
\keyword{internal}