% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_backend.R
\name{set_backend}
\alias{set_backend}
\title{Set the backend for the chopin package}
\usage{
set_backend(backend = "mirai")
}
\arguments{
\item{backend}{The backend to be set. Default is "mirai".}
}
\value{
None
}
\description{
This function sets the backend for the chopin package to be used for computation.
}
