% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_computational_domain.R
\name{sp_index_grid}
\alias{sp_index_grid}
\title{sp_index_grid: Generate grid polygons}
\usage{
sp_index_grid(points_in, ncutsx, ncutsy)
}
\arguments{
\item{points_in}{sf or SpatVector object. Target points of computation.}

\item{ncutsx}{integer(1). The number of splits along x-axis.}

\item{ncutsy}{integer(1). The number of splits along y-axis.}
}
\value{
A sf or SpatVector object of computation grids with unique grid id (CGRIDID).
}
\description{
Returns a sf object that includes x- and y- index by using two inputs ncutsx and ncutsy, which are x- and y-directional splits, respectively.
}
\author{
Insang Song
}
