% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{sp_indexing}
\alias{sp_indexing}
\title{Create integer indices for grid}
\usage{
sp_indexing(points_in, ncutsx, ncutsy)
}
\arguments{
\item{points_in}{sf object.}

\item{ncutsx}{integer(1). The number of splits along x-axis.}

\item{ncutsy}{integer(1). The number of splits along y-axis.}
}
\description{
Returns a tibble object that includes x- and y- index by using two inputs ncutsx and ncutsy, which are x- and y-directional splits, respectively.
}
\author{
Insang Song
}
