% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{summarize_aw_old}
\alias{summarize_aw_old}
\title{Area weighted summary using two polygon sf or SpatVector objects
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}}
\usage{
summarize_aw_old(
  poly_in = NULL,
  poly_weight = NULL,
  target_fields = NULL,
  id_poly_in = "ID",
  fun = stats::weighted.mean,
  extent = NULL
)
}
\arguments{
\item{poly_in}{A sf/SpatVector object or file path of polygons detectable
with GDAL driver at weighted means will be calculated.}

\item{poly_weight}{A sf/SpatVector object or file path of polygons from
which weighted means will be calculated.}

\item{target_fields}{character. Field names to calculate area-weighted.}

\item{id_poly_in}{character(1).
The unique identifier of each polygon in \code{poly_in}.
Default is \code{"ID"}.}

\item{fun}{function(1). The function to calculate the weighted summary.
Default is \code{\link[stats:weighted.mean]{stats::weighted.mean}}. The function must have a \code{w} argument.}

\item{extent}{numeric(4) or SpatExtent object. Extent of clipping \code{poly_in}.
It only works with \code{poly_in} of character(1) file path.
See \code{\link[terra:ext]{terra::ext}} for more details. Coordinate systems should match.}
}
\value{
A data.frame with all numeric fields of area-weighted means.
}
\description{
When \code{poly_in} and \code{poly_weight} are different classes,
\code{poly_weight} will be converted to the class of \code{poly_in}.
}
\note{
If \code{poly_in} and \code{poly_weight} are characters, they will be
read as \code{terra::vect} objects.
}
\seealso{
Other Macros for calculation: 
\code{\link{extract_at}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_aw}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}
