% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{summarize_pp}
\alias{summarize_pp}
\alias{summarize_pp,SpatVector-method}
\alias{summarize_pp,character-method}
\alias{summarize_pp,sf-method}
\title{Point to polygon summary using target polygons and source points}
\usage{
summarize_pp(x, y, ...)

\S4method{summarize_pp}{SpatVector}(
  x,
  y,
  target_fields = NULL,
  id_x = "ID",
  fun = mean,
  extent = NULL,
  ...
)

\S4method{summarize_pp}{character}(
  x,
  y,
  target_fields = NULL,
  id_x = "ID",
  fun = mean,
  out_class = "terra",
  extent = NULL,
  ...
)

\S4method{summarize_pp}{sf}(
  x,
  y,
  target_fields = NULL,
  id_x = "ID",
  fun = mean,
  extent = NULL,
  ...
)
}
\arguments{
\item{x}{A sf/SpatVector object or file path of polygons to which
point attributes will be summarized.}

\item{y}{A sf/SpatVector object or file path of points from
which summaries will be calculated.}

\item{...}{Additional arguments passed to \code{fun}.}

\item{target_fields}{character. Field names in \code{y} to summarize.}

\item{id_x}{character(1).
The unique identifier of each polygon in \code{x}.
Default is \code{"ID"}.}

\item{fun}{function(1)/character(1).
The function to calculate the point summary.
Default is \code{\link{mean}}. Character input should be a summary
function accepted by \code{match.fun()}.}

\item{extent}{numeric(4) or SpatExtent object. Extent of clipping \code{x}.
It only works with \code{x} of character(1) file path.
See \code{\link[terra:ext]{terra::ext}} for more details. Coordinate systems should match.}

\item{out_class}{character(1). "sf" or "terra". Output class.}
}
\value{
If \code{x} is \code{sf}, returns an \code{sf} object with summarized fields
joined to polygons. Otherwise returns a \code{data.frame}.
}
\description{
When \code{x} and \code{y} are different classes,
\code{y} will be converted to the class of \code{x}.
}
\note{
\code{x} should contain polygon geometries and \code{y} should contain
point geometries.
}
\seealso{
Other Macros for calculation:
\code{\link[=extract_at]{extract_at()}},
\code{\link[=kernelfunction]{kernelfunction()}},
\code{\link[=summarize_aw]{summarize_aw()}},
\code{\link[=summarize_sedc]{summarize_sedc()}},
\code{\link[=summarize_st]{summarize_st()}}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}
