% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{usher}
\alias{usher}
\title{Usher data ingestion}
\usage{
usher(x)
}
\arguments{
\item{x}{character, data.frame, any object.}
}
\description{
The input will be inspected to determine the workflow of the data
ingestion.
}
