% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{validate_and_repair_vectors}
\alias{validate_and_repair_vectors}
\title{Validate and repair input vector data}
\usage{
validate_and_repair_vectors(input_vector)
}
\arguments{
\item{input_vector}{One of sf or vect class. Target points of computation.}
}
\value{
A repaired sf or SpatVector object depending on the class of input_vector.
}
\description{
It tries repairing input vector data. Vector validity violation usually appears in polygon data with self-crossing or hole orders. This function will pass the input_vector object to sf::st_make_valid() (if input_vector is sf) or terra::makeValid() (if input_vector is SpatVector). May take some time depending on the geometry complexity.
}
\author{
Insang Song
}