% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{vect_valid_repair}
\alias{vect_valid_repair}
\title{Validate and repair input vector data}
\usage{
vect_valid_repair(input_vector)
}
\arguments{
\item{input_vector}{One of sf or vect class. Target points of computation.}
}
\value{
A repaired \code{sf} or \code{SpatVector} object depending on
the class of input_vector.
}
\description{
It tries repairing input vector data.
Vector validity violation usually appears in polygon data with
self-crossing or
hole orders. This function will pass the input_vector object to
\code{\link[sf:valid]{sf::st_make_valid}} (if input_vector is sf) or
\code{\link[terra:is.valid]{terra::makeValid}} (if input_vector is SpatVector).
May take some time depending on the geometry complexity.
}
\examples{
\dontrun{
library(terra)
library(sf)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)

nc_valid <- vect_valid_repair(nc)
}
}
\author{
Insang Song
}