% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{vect_validate}
\alias{vect_validate}
\title{Validate and repair input vector data}
\usage{
vect_validate(input_vector)
}
\arguments{
\item{input_vector}{One of sf or vect class. Target points of computation.}
}
\value{
A repaired \code{sf} or \code{SpatVector} object depending on
the class of input_vector.
}
\description{
It tries repairing input vector data.
Vector validity violation usually appears in polygon data with
self-crossing or
hole orders. This function will pass the input_vector object to
\code{\link[sf:valid]{sf::st_make_valid}} (if input_vector is sf) or
\code{\link[terra:is.valid]{terra::makeValid}} (if input_vector is SpatVector).
May take some time depending on the geometry complexity.
}
\note{
This function works with GEOS (>=3.8).
}
\examples{
\dontrun{
library(terra)
library(sf)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)

nc_valid <- vect_validate(nc)
}
}
\seealso{
Other Helper functions: 
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{par_def_q}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
\keyword{soft-deprecated}
