% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{any_class_args}
\alias{any_class_args}
\title{Detect classes in function arguments}
\usage{
any_class_args(args = NULL, search = NULL)
}
\arguments{
\item{args}{Any list, but preferably generated by \code{list(...)} inside
a function.}

\item{search}{character(1). Class name to search. Partial match is supported.}
}
\value{
logical vector.
}
\description{
When a R function is defined in an ordinary
fashion (i.e., assigning a function by \code{<- function(...)})
would be subject to ambiguity particularly if the function
name is the same as the generic function name(s).
This function supports detecting classes of arguments in
a loosely defined function.
}
\examples{
df <- data.frame(a = 1, b = 3)
any_class_args(list(df), "data.frame")
}
\seealso{
Other Helper functions: 
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{par_def_q}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}
