% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_bbox}
\alias{check_bbox}
\title{Check if the data extent is inside the reference bounding box}
\usage{
check_bbox(data_query, reference)
}
\arguments{
\item{data_query}{sf*/stars/SpatVector/SpatRaster object.}

\item{reference}{sf*/stars/SpatVector/SpatRaster object}
}
\value{
TRUE (the queried data extent is completely within
the reference bounding box) or FALSE
}
\description{
One of the most common errors in
spatial computation is rooted in
the entirely or partly incomparable spatial extents of input datasets.
This function returns whether your data is inside
the target computational extent.
It is assumed that you know and have the exact computational region.
This function will return TRUE if the reference region
completely contains your data's extent and FALSE otherwise.
}
\author{
Insang Song \email{geoissong@gmail.com}
}