% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{check_datatype}
\alias{check_datatype}
\title{Return the data type}
\usage{
check_datatype(input)
}
\arguments{
\item{input}{Spat*/sf/stars object.}
}
\value{
character(1). One of 'vector' or 'raster'.
}
\description{
This function returns one of 'vector' or 'raster' depending on the input class.
}
\note{
\code{stars} object is tricky a little whether to classify vector or raster. In this package, \code{stars} object will be considered raster.
}
\author{
Insang song
}