% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dist_incorrect}
\alias{check_dist_incorrect}
\title{Check the distance calculated at a unit grid would be suspicious}
\usage{
check_dist_incorrect(points, grid, dist_calc)
}
\arguments{
\item{points}{SpatVector point object.}

\item{grid}{plain grid (i.e., not padded from \code{\link{par_pad_grid}})}

\item{dist_calc}{Calculated distance.}
}
\value{
logical vector.
}
\description{
This function checks if the alphahull points reported
the distance to the nearest computational grid boundary.
}
\details{
Alpha hull refers to the parametrized exterior boundary
of a set of points. This function checks if the distance to the nearest
computational grid boundary is reported in the alphahull points.
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{par_def_q}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\concept{Helper functions}
