% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_within_reference}
\alias{check_within_reference}
\title{Check if the boundary of the vector/raster object is inside the reference}
\usage{
check_within_reference(input_object, reference)
}
\arguments{
\item{input_object}{sf/stars/SpatVector/SpatRaster object.}

\item{reference}{sf/stars/SpatVector/SpatRaster object.}
}
\value{
logical
}
\description{
Check if the boundary of the vector/raster object is inside the reference
}
\author{
Insang Song \email{geoissong@gmail.com}
}
