% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopin-package.R
\name{chopin-package}
\alias{chopin-package}
\alias{chopin}
\title{Computation of spatial data by hierarchical and objective partitioning of inputs for parallel processing}
\description{
The \code{chopin} package provides a set of functions to compute on divided
geospatial data.
}
\keyword{internal}
