% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopin-package.R
\docType{package}
\name{chopin-package}
\alias{chopin-package}
\alias{chopin}
\title{Computation of spatial data by hierarchical and objective partitioning of inputs for parallel processing}
\description{
The \code{chopin} package provides a set of functions to compute on divided
geospatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIEHS/chopin}
  \item Report bugs at \url{https://github.com/NIEHS/chopin/issues}
}

}
\author{
\strong{Maintainer}: Insang Song \email{geoissong@gmail.com} (\href{https://orcid.org/0000-0001-8732-3256}{ORCID})

Authors:
\itemize{
  \item Kyle Messier (\href{https://orcid.org/0000-0001-9508-9623}{ORCID}) [contributor]
}

}
\keyword{internal}
