% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{clip_as_extent_ras}
\alias{clip_as_extent_ras}
\title{clip_as_extent_ras: Clip input raster.}
\usage{
clip_as_extent_ras(pnts, buffer_r, ras, nqsegs = 180L)
}
\arguments{
\item{pnts}{sf or SpatVector object}

\item{buffer_r}{numeric(1). buffer radius.
This value will be automatically multiplied by 1.25}

\item{ras}{SpatRaster object to be clipped}

\item{nqsegs}{integer(1). the number of points per a quarter circle}
}
\description{
Clip input raster by the expected maximum extent of computation.
}
\author{
Insang Song
}