% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{clip_ras_ext}
\alias{clip_ras_ext}
\title{Clip input raster with a buffered vector extent.}
\usage{
clip_ras_ext(pnts = NULL, radius = NULL, ras = NULL, nqsegs = 180L)
}
\arguments{
\item{pnts}{\code{sf} or \code{SpatVector} object}

\item{radius}{numeric(1). buffer radius.
This value will be automatically multiplied by 1.25}

\item{ras}{\code{SpatRaster} object to be clipped}

\item{nqsegs}{\code{integer(1)}. the number of points per a quarter circle}
}
\description{
Clip input raster by the expected maximum extent of
computation.
}
\examples{
library(terra)

ras_rand <- terra::rast(nrow = 20, ncol = 20)
terra::values(ras_rand) <- runif(400L)
ras_rand_p <-
  data.frame(
    x = c(3, 5, 3.2, 8),
    y = c(12, 10, 15, 12),
    z = c(0, 1, 2, 3)
  )
ras_rand_p <- terra::vect(ras_rand_p, geom = c("x", "y"))
clip_ras_ext(ras_rand_p, 1.5, ras_rand)
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext2poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{rast_short}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}