% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clip_vec_ext}
\alias{clip_vec_ext}
\title{Clip to the buffered extent of input vector}
\usage{
clip_vec_ext(x, y, radius)
}
\arguments{
\item{x}{\code{sf} or \code{SpatVector} object to be clipped}

\item{y}{\code{sf} or \code{SpatVector} object}

\item{radius}{\code{numeric(1)}. Circular buffer radius.
this value will be automatically multiplied by 1.1}
}
\value{
A clipped \code{sf} or \code{SpatVector} object.
}
\description{
Clip input vector by
the expected maximum extent of computation.
}
\examples{
library(sf)
library(stars)
library(terra)
options(sf_use_s2 = FALSE)

bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
bcsd <- stars::read_stars(bcsd_path)
bcsd <- sf::st_as_sf(bcsd)
bcsd_rpnt <- sf::st_as_sf(sf::st_sample(bcsd, 4L))
bcsd_rpntm <- sf::st_as_sf(sf::st_sample(bcsd, 1000L))
clip_vec_ext(bcsd_rpntm, 1000, bcsd_rpnt)
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{par_def_q}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
