% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{crs_check}
\alias{crs_check}
\title{Check Coordinate Reference System}
\usage{
crs_check(x)
}
\arguments{
\item{x}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object.}
}
\value{
A st_crs or crs object.
}
\description{
It returns st_crs object from \code{sf}/Spat* objects.
}
\examples{
# data
library(sf)
ncpath = system.file("shape/nc.shp", package = "sf")
nc = read_sf(ncpath)
crs_check(nc)
}
\author{
Insang Song \email{geoissong@gmail.com}
}