% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{crs_check}
\alias{crs_check}
\title{Check Coordinate Reference System}
\usage{
crs_check(x = NULL)
}
\arguments{
\item{x}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object.}
}
\value{
A st_crs or crs object.
}
\description{
It returns st_crs object from \code{sf}/Spat* objects.
}
\examples{
# data
library(sf)
ncpath = system.file("shape/nc.shp", package = "sf")
nc = read_sf(ncpath)
crs_check(nc)
}
\seealso{
Other Helper functions: 
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{par_def_q}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Helper functions}
\keyword{internal}
\keyword{soft-deprecated}
