% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{datamod}
\alias{datamod}
\title{Return the input's GIS data model type}
\usage{
datamod(input)
}
\arguments{
\item{input}{Spat*/sf/stars object.}
}
\value{
character(1). One of \code{"vector"} or \code{"raster"}.
}
\description{
This function returns one of 'vector' or 'raster'
depending on the input class.
}
\note{
Although \code{stars} object is a little ambiguous
whether to classify vector or raster,
it will be considered raster in this package.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
datamod(nc_sf)

ra_path <- system.file("ex/elev.tif", package = "terra")
ra <- terra::rast(ra_path)
datamod(ra)
}
\author{
Insang Song
}