% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dep_check}
\alias{dep_check}
\title{Return the package the input object is based on}
\usage{
dep_check(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
A character object; one of \code{"character"}, \code{"terra"} and \code{"sf"}
}
\description{
Detect whether the input object is sf or Spat* object.
}
\seealso{
Other Helper functions: 
\code{\link{datamod}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{par_def_q}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
