% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dep_check}
\alias{dep_check}
\title{Return the package the input object is based on}
\usage{
dep_check(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
A character object; one of \code{"terra"} and \code{"sf"}
}
\description{
Detect whether the input object is sf or Spat* object.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
dep_check(nc_sf)
nc_vect <- terra::vect(nc_sf)
dep_check(nc_vect)
## END OF EXAMPLE
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_switch}()},
\code{\link{ext2poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{par_def_q}()},
\code{\link{reproject_b2r}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}
