% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dep_switch}
\alias{dep_switch}
\title{Switch spatial data class}
\usage{
dep_switch(input)
}
\arguments{
\item{input}{Spat* in terra or sf object.}
}
\value{
Data converted to the other package class
(if sf, terra; if terra, sf)
}
\description{
Convert class between \code{sf}/\code{stars}-\code{terra}
}
\examples{
library(sf)
library(stars)
library(terra)
options(sf_use_s2 = FALSE)

## generate a random raster
ras_rand <- terra::rast(nrow = 30, ncol = 30)
terra::values(ras_rand) <- runif(900)
stars_rand <- dep_switch(ras_rand)
stars_rand
# should return stars object

vec_rand <- terra::spatSample(ras_rand, size = 10L, as.points = TRUE)
sf_rand <- dep_switch(vec_rand)
sf_rand
# should return sf object
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_within_ref}()},
\code{\link{par_def_q}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
