% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.check_subject}
\alias{.check_subject}
\title{Check the subject object and perform necessary conversions if needed.}
\usage{
.check_subject(subject, subject_id, extent = NULL, out_class = "sf")
}
\arguments{
\item{subject}{sf/SpatVector/character. The input object to be checked.}

\item{subject_id}{character(1). ID field of the subject object.}

\item{extent}{numeric(4). The extent of the subject object.
Numeric vector should be put in order of
\code{c(xmin, xmax, ymin, ymax)}.}

\item{out_class}{character(1). The class of the output object.
Should be one of \code{c("sf", "terra")}.}
}
\value{
The checked and converted subject object.
}
\description{
This function checks the class of the input object and
performs necessary conversions if needed.
}
\examples{
# Check a SpatVector object
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
extent <- c(-80, -77, 35, 36)
.check_subject(subject = nc, extent = extent, subject_id = "FIPS")

# Check a sf object
ncsf <- sf::st_read(ncpath)
.check_subject(subject = ncsf, extent = extent, subject_id = "FIPS")

# Check a character object
.check_subject(
  subject = ncpath,
  extent = extent,
  out_class = "terra",
  subject_id = "FIPS"
)
}
\keyword{internal}
