% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.check_subject}
\alias{.check_subject}
\title{Check the subject object and perform necessary conversions if needed.}
\usage{
.check_subject(input, input_id = NULL, extent = NULL, out_class = "sf", ...)
}
\arguments{
\item{input}{sf/SpatVector/character. The input object to be checked.}

\item{input_id}{character(1). ID field of the subject object.}

\item{extent}{numeric(4). The extent of the subject object.
Numeric vector should be put in order of
\code{c(xmin, xmax, ymin, ymax)}.}

\item{out_class}{character(1). The class of the output object.
Should be one of \code{c("sf", "terra")}.}

\item{...}{Placeholder.}
}
\value{
The checked and converted subject object.
}
\description{
This function checks the class of the input object and
performs necessary conversions if needed.
}
\examples{
# Check a SpatVector object
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
extent <- c(-80, -77, 35, 36)
.check_subject(input = nc, extent = extent, input_id = "FIPS")

# Check a sf object
ncsf <- sf::st_read(ncpath)
.check_subject(input = ncsf, extent = extent, input_id = "FIPS")

# Check a character object
.check_subject(
  input = ncpath,
  extent = extent,
  out_class = "terra",
  input_id = "FIPS"
)
}
\keyword{internal}
