% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{extract_at}
\alias{extract_at}
\title{Extract raster values with point buffers or polygons}
\usage{
extract_at(
  vector = NULL,
  raster = NULL,
  id = NULL,
  func = "mean",
  mode = c("polygon", "buffer"),
  ...
)
}
\arguments{
\item{vector}{\code{sf}/\code{SpatVector} object.}

\item{raster}{\code{SpatRaster} object. or file path(s) with extensions
that are GDAL-compatible.}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{function taking one numeric vector argument.}

\item{mode}{one of \code{"polygon"}
(generic polygons to extract raster values with) or
\code{"buffer"} (point with buffer radius)}

\item{...}{various. Passed to extract_at_buffer.
See \code{?extract_at_buffer} for details.}
}
\value{
A data.frame object with summarized raster values with
respect to the mode (polygon or buffer) and the function.
}
\description{
Extract raster values with point buffers or polygons
}
\examples{
## See ?extract_at_poly and ?extract_at_buffer
}
\seealso{
\link{extract_at_poly}, \link{extract_at_buffer}

Other Macros for calculation: 
\code{\link{extract_at_buffer}()},
\code{\link{extract_at_poly}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_aw}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}