% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{extract_at_poly}
\alias{extract_at_poly}
\title{Extract summarized values from raster with generic polygons}
\usage{
extract_at_poly(polys = NULL, surf = NULL, id = NULL, func = "mean")
}
\arguments{
\item{polys}{\code{sf}/\code{SpatVector} object. Polygons.}

\item{surf}{\code{SpatRaster} object.
A raster from which a summary will be calculated}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{a generic function name in string or
a function taking two arguments that are
compatible with \code{\link[exactextractr]{exact_extract}}.
For example, \code{"mean"} or \verb{\\(x, w) weighted.mean(x, w, na.rm = TRUE)}}
}
\value{
a data.frame object with function value
}
\description{
For simplicity, it is assumed that the coordinate systems of
the points and the raster are the same.
Kernel function is not yet implemented.
}
\examples{
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
nc <- terra::project(nc, "EPSG:5070")
rrast <- terra::rast(nc, nrow = 100, ncol = 220)
ncr <- terra::rasterize(nc, rrast)
terra::values(rrast) <- rgamma(2.2e4, 4, 2)
rpnt <- terra::spatSample(rrast, 16L, as.points = TRUE)
rpnt$pid <- sprintf("ID-\%02d", seq(1, 16))
rpoly <-
  terra::buffer(rpnt, 5, capstyle = "square", joinstyle = "bevel")
extract_at_poly(rpoly, rrast, "pid")
}
\seealso{
Other Macros for calculation: 
\code{\link{extract_at}()},
\code{\link{extract_at_buffer}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_aw}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}