% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_with}
\alias{extract_with}
\title{Extract raster values with point buffers or polygons}
\usage{
extract_with(
  raster,
  vector,
  id,
  func = mean,
  mode = c("polygon", "buffer"),
  ...
)
}
\arguments{
\item{raster}{SpatRaster object.}

\item{vector}{SpatVector object.}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{function taking one numeric vector argument.}

\item{mode}{one of "polygon" (generic polygons to extract raster values with) or "buffer" (point with buffer radius)}

\item{...}{various. Passed to extract_with_buffer. See \code{?extract_with_buffer} for details.}
}
\description{
Extract raster values with point buffers or polygons
}
\author{
Insang Song \email{geoissong@gmail.com}
}
