% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{extract_with_polygons}
\alias{extract_with_polygons}
\title{Extract summarized values from raster with generic polygons}
\usage{
extract_with_polygons(polys, surf, id, func = mean, na.rm = TRUE)
}
\arguments{
\item{polys}{sf/SpatVector object. Polygons.}

\item{surf}{stars/SpatRaster object. A raster of whatnot a summary will be calculated}

\item{id}{character(1). Unique identifier of each point.}

\item{func}{a function taking one argument. For example, function(x) mean(x, na.rm = TRUE) or \(x) mode(x, na.rm = TRUE)}

\item{na.rm}{logical(1). NA values are omitted when summary is calculated.}
}
\value{
a data.frame object with function value
}
\description{
For simplicity, it is assumed that the coordinate systems of the points and the raster are the same. Kernel function is not yet implemented.
}
\author{
Insang Song \email{geoissong@gmail.com}
}