% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{get_clip_ext}
\alias{get_clip_ext}
\title{Setting the clipping extent}
\usage{
get_clip_ext(pnts, radius)
}
\arguments{
\item{pnts}{One of sf or SpatVector object. Target points of computation.}

\item{radius}{numeric(1). Buffer radius. It is assumed to be in meters}
}
\value{
A terra::ext or sfc_POLYGON object of the computation extent.
}
\description{
Return clipping extent with buffer radius.
It assumes the input CRS is projected and linear unit is meters.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
get_clip_ext(nc_sf)
nc_vect <- terra::vect(nc_sf)
get_clip_ext(nc_vect)
}
\author{
Insang Song
}