% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{kernelfunction}
\alias{kernelfunction}
\title{Kernel functions}
\usage{
kernelfunction(
  d,
  bw,
  kernel = c("uniform", "quartic", "triweight", "epanechnikov")
)
}
\arguments{
\item{d}{Distance}

\item{bw}{Bandwidth of a kernel}

\item{kernel}{Kernel type. One of
'uniform', 'quartic', 'triweight', and 'epanechnikov'}
}
\description{
Kernel functions
}
\references{
\href{https://github.com/JanCaha/SpatialKDE}{SpatialKDE source}
}