% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ncpoints}
\alias{ncpoints}
\title{Mildly clustered points in North Carolina, United States}
\format{
A data frame with 2,304 rows and two variables:
\describe{
\item{X}{X coordinate}
\item{Y}{Y coordinate}
}
}
\source{
sf package data \code{nc}
}
\usage{
ncpoints
}
\description{
Mildly clustered points in North Carolina, United States
}
\note{
Coordinates are in EPSG:5070 (Conus Albers Equal Area)
}
\examples{
data("ncpoints", package = "chopin")
}
\seealso{
Other Dataset: 
\code{\link{prediction_grid}}
}
\concept{Dataset}
\keyword{datasets}
