% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridding.R
\name{par_cut_coords}
\alias{par_cut_coords}
\title{Partition coordinates into quantile polygons}
\usage{
par_cut_coords(x = NULL, y = NULL, quantiles)
}
\arguments{
\item{x}{numeric/sf/SpatVector. x-coordinates (if numeric).}

\item{y}{numeric. y-coordinates.}

\item{quantiles}{numeric vector. Quantiles.}
}
\value{
A \code{SpatVector} object with field \code{CGRIDID}.
}
\description{
Partition coordinates into quantile polygons
}
\note{
This function is only for two-dimensional points.
}
\examples{
library(terra)
random_points <-
  data.frame(x = runif(1000, 0, 100), y = runif(1000, 0, 100))
quantiles <- seq(0, 1, length.out = 5L)
qpoly <- par_cut_coords(random_points$x, random_points$y, quantiles)
clustered_points <-
  data.frame(x = rgamma(1000, 1, 1), y = rgamma(1000, 4, 1))
qpoly_c <- par_cut_coords(clustered_points$x, clustered_points$y, quantiles)
par(mfcol = c(1, 2))
plot(qpoly)
plot(qpoly_c)
par(mfcol = c(1, 1))
cvect <- terra::vect(clustered_points, geom = c("x", "y"))
plot(cvect)
plot(qpoly_c, add = TRUE, col = "transparent", border = "red")
qcv <- intersect(cvect, qpoly_c)
table(qcv$CGRIDID)
sum(table(qcv$CGRIDID)) # should be 1000
}
\seealso{
Other Parallelization: 
\code{\link{par_grid}()},
\code{\link{par_hierarchy}()},
\code{\link{par_make_grid}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()},
\code{\link{par_pad_balanced}()},
\code{\link{par_pad_grid}()},
\code{\link{par_split_list}()}
}
\concept{Parallelization}
\keyword{internal}
