% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridding.R
\name{par_def_q}
\alias{par_def_q}
\title{Quantile definition}
\usage{
par_def_q(steps = 4L)
}
\arguments{
\item{steps}{integer(1). The number of quantiles.}
}
\value{
numeric vector of quantiles.
}
\description{
Quantile definition
}
\examples{
par_def_q(5L)
}
\seealso{
Other Helper functions: 
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\concept{Helper functions}
\keyword{internal}
