% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R
\name{par_fallback}
\alias{par_fallback}
\title{Parallelization error fallback}
\usage{
par_fallback(err = NULL, fun = NULL, debug = FALSE)
}
\arguments{
\item{err}{Error status or message.}

\item{fun}{function.}

\item{debug}{logical(1). Print error messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
data.frame with one column
}
\description{
Parallelization error fallback
}
\note{
This function assumes that the \code{fun} has an argument named
\code{"id"}.
}
\examples{
err <- simpleError("No input.")
par_fallback(err, extract_at, debug = TRUE)
}
\seealso{
Other Parallelization: 
\code{\link{par_cut_coords}()},
\code{\link{par_grid}()},
\code{\link{par_group_grid}()},
\code{\link{par_hierarchy}()},
\code{\link{par_make_grid}()},
\code{\link{par_make_gridset}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()}
}
\author{
Insang Song
}
\concept{Parallelization}
