% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{par_fallback}
\alias{par_fallback}
\title{Parallelization error fallback}
\usage{
par_fallback(err = NULL, fun = NULL, debug = FALSE)
}
\arguments{
\item{err}{Error status or message.}

\item{fun}{function.}

\item{debug}{logical(1). Print error messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
data.frame with one column
}
\description{
Parallelization error fallback
}
\note{
This function assumes that the \code{fun} has an argument named
\code{"id"}.
}
\examples{
## NO EXAMPLE
}
\author{
Insang Song
}