% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R
\name{par_grid}
\alias{par_grid}
\title{Process a given function in the entire or partial computational grids}
\usage{
par_grid(
  grids,
  grid_target_id = NULL,
  debug = FALSE,
  combine = collapse::rowbind,
  fun_dist,
  ...
)
}
\arguments{
\item{grids}{sf/SpatVector object. Computational grids.
It takes a strict assumption that the grid input is
an output of \code{par_make_gridset}.
If missing or NULL is entered, \code{par_make_gridset} is internally
called. In this case, the \strong{first} element of the ellipsis argument
\code{...} is considered \code{input} in \code{par_make_gridset} and \code{mode} is fixed
as "grid"`. See \code{\link[=par_make_gridset]{par_make_gridset()}} for details.}

\item{grid_target_id}{character(1) or numeric(2).
Default is NULL. If NULL, all grid_ids are used.
\code{"id_from:id_to"} format or
\code{c(unique(grid_id)[id_from], unique(grid_id)[id_to])}}

\item{debug}{logical(1). Prints error messages
if there were any errors during the calculation.}

\item{combine}{function. A function to combine results.
Default is \code{dplyr::bind_rows}.}

\item{fun_dist}{\code{sf}, \code{terra} or \code{chopin} functions.}

\item{...}{Arguments passed to the argument \code{fun_dist}.
The \strong{second} place should get a vector or raster dataset from which
you want to extract or calculate values. For example, a raster dataset
when vector-raster overlay is performed.}
}
\value{
a data.frame object with computation results.
For entries of the results, consult the function used in
\code{fun_dist} argument.
}
\description{
\link[future:multicore]{future::multicore}, \link[future:multisession]{future::multisession}, \link[future:cluster]{future::cluster}
with \link[doParallel:registerDoParallel]{doParallel::registerDoParallel} will parallelize the work
in each grid. For details of the terminology in \code{future} package,
refer to \link[future]{plan}. This function assumes that
users have one raster file and a sizable and spatially distributed
target locations. Each thread will process
the nearest integer of $|N_g| / |N_t|$ grids
where $|N_g|$ denotes the number of grids and $|N_t|$ denotes
the number of threads.
}
\note{
In dynamic dots (\code{...}), the first and second
arguments should be the \code{fun_dist} arguments where
sf/SpatVector objects are accepted.
Virtually any sf/terra functions that accept two arguments
can be put in \code{fun_dist}, but please be advised that
some spatial operations do not necessarily give the
exact result from what would have been done single-thread.
For example, distance calculated through this function may return the
lower value than actual because the computational region was reduced.
This would be the case especially where the target features
are spatially sparsely distributed.
}
\examples{
\dontrun{
ncpath <- system.file("shape/nc.shp", package = "sf")
ncpoly <- terra::vect(ncpath) |>
  terra::project("EPSG:5070")
ncpnts <-
  readRDS(
    system.file("extdata/nc_random_point.rds", package = "chopin")
  )
ncpnts <- terra::vect(ncpnts)
ncpnts <- terra::project(ncpnts, "EPSG:5070")
ncelev <-
  terra::unwrap(
    readRDS(system.file("extdata/nc_srtm15_otm.rds", package = "chopin"))
  )
terra::crs(ncelev) <- "EPSG:5070"
names(ncelev) <- c("srtm15")

ncsamp <-
  terra::spatSample(
    terra::ext(ncelev),
    1e4L,
    lonlat = FALSE,
    as.points = TRUE
  )
ncsamp$kid <- sprintf("K-\%05d", seq(1, nrow(ncsamp)))
nccompreg <-
  par_make_gridset(
    input = ncpnts,
    mode = "grid",
    nx = 6L,
    ny = 4L,
    padding = 3e4L
  )
res <-
  par_grid(
    grids = nccompreg,
    grid_target_id = NULL,
    fun_dist = extract_at_buffer,
    points = ncpnts,
    surf = ncelev,
    qsegs = 90L,
    radius = 5e3L,
    id = "pid"
  )
}
}
\seealso{
Other Parallelization: 
\code{\link{par_cut_coords}()},
\code{\link{par_fallback}()},
\code{\link{par_group_grid}()},
\code{\link{par_hierarchy}()},
\code{\link{par_make_grid}()},
\code{\link{par_make_gridset}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Parallelization}
