% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R
\name{par_hierarchy}
\alias{par_hierarchy}
\title{Process a given function using a hierarchy in input data}
\usage{
par_hierarchy(
  regions,
  regions_id = NULL,
  fun_dist,
  ...,
  pad_y = FALSE,
  .debug = FALSE
)
}
\arguments{
\item{regions}{\code{sf}/\code{SpatVector} object.
Computational regions. Only polygons are accepted.}

\item{regions_id}{character(nrow(regions)) or character(1).
The regions will be split by the common level value.
The level should be higher than the original data level.
A field name with the higher level information is also accepted.}

\item{fun_dist}{\code{sf}, \code{terra}, or \code{chopin} functions.
This function should have \code{x} and \code{y} arguments.}

\item{...}{Arguments passed to the argument \code{fun_dist}.
The \strong{second} place should get a vector or raster dataset from which
you want to extract or calculate values. For example, a raster dataset
when vector-raster overlay is performed.}

\item{pad_y}{logical(1). Whether to filter y with the padded grid.
Should be TRUE when x is where the values are calculated.
Default is \code{FALSE}. In the flipped case, like \code{terra::extent} or
\code{exactextractr::exact_extract}, the raster (x) should be scoped
with the padded grid.}

\item{.debug}{logical(1). Default is \code{FALSE}
If a unit computation fails, the error message and the \code{regions_id}
value where the error occurred will be included in the output.}
}
\value{
a data.frame object with computation results.
For entries of the results, consult the function used in
\code{fun_dist} argument.
}
\description{
"Hierarchy" refers to a system,
which divides the entire study region into multiple subregions.
It is oftentimes reflected in an area code system
(e.g., FIPS for US Census geographies, HUC-4, -6, -8, etc.).
\code{\link[future:multisession]{future::multisession}}, \code{\link[future:multicore]{future::multicore}}, \code{\link[future:cluster]{future::cluster}},
\code{\link[future.mirai:mirai_multisession]{future.mirai::mirai_multisession}} in \code{\link[future:plan]{future::plan}}
will parallelize the work by splitting lower level features into
several higher level feature group.
For details of the terminology in \code{future} package,
refer to \code{\link[future:plan]{future::plan}}.
Each thread will process the number of lower level features
in each higher level feature. Please be advised that
accessing the same file simultaneously with
multiple processes may result in errors.
}
\note{
In dynamic dots (\code{...}), \code{fun_dist} arguments should include
x and y where sf/terra class objects or file paths are accepted.
Virtually any sf/terra functions that accept two arguments
can be put in \code{fun_dist}, but please be advised that
some spatial operations do not necessarily give the
exact result from what would have been done single-thread.
For example, distance calculated through this function may return the
lower value than actual because the computational region was reduced.
This would be the case especially where the target features
are spatially sparsely distributed.
}
\examples{
library(terra)
library(sf)
library(future)
library(future.mirai)
sf::sf_use_s2(FALSE)
future::plan(future.mirai::mirai_multisession, workers = 2)

ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
nccnty <- sf::st_read(ncpath, layer = "county")
nctrct <- sf::st_read(ncpath, layer = "tracts")
ncelev <-
  system.file("extdata/nc_srtm15_otm.tif", package = "chopin")

ncsamp <-
  sf::st_sample(
    nccnty,
    size = 1e4L
  )
ncsamp <- sf::st_as_sf(ncsamp)
ncsamp$kid <- sprintf("K-\%05d", seq_len(nrow(ncsamp)))
res <-
  par_hierarchy(
    regions = nccnty,
    regions_id = "GEOID",
    fun_dist = extract_at,
    y = nctrct,
    x = ncelev,
    id = "GEOID",
    func = "mean"
  )
)
}
\seealso{
\code{\link[future:multisession]{future::multisession}} \code{\link[future:multicore]{future::multicore}} \code{\link[future:cluster]{future::cluster}}
\code{\link[future.mirai:mirai_multisession]{future.mirai::mirai_multisession}} \code{\link[future:plan]{future::plan}} \code{\link{par_map_args}}

Other Parallelization: 
\code{\link{par_cut_coords}()},
\code{\link{par_grid}()},
\code{\link{par_make_grid}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()},
\code{\link{par_pad_balanced}()},
\code{\link{par_pad_grid}()},
\code{\link{par_split_list}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Parallelization}
