% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridding.R
\name{par_make_grid}
\alias{par_make_grid}
\title{Generate grid polygons}
\usage{
par_make_grid(points_in = NULL, ncutsx = NULL, ncutsy = NULL)
}
\arguments{
\item{points_in}{\code{sf} or \code{SpatVector} object. Target points of computation.}

\item{ncutsx}{integer(1). The number of splits along x-axis.}

\item{ncutsy}{integer(1). The number of splits along y-axis.}
}
\value{
A \code{sf} or \code{SpatVector} object of computation grids with
unique grid id (CGRIDID).
}
\description{
Returns a sf object that includes x- and y- index
by using two inputs ncutsx and ncutsy, which are x- and
y-directional splits, respectively.
}
\note{
Grids are generated based on the extent of \code{points_in} first,
then exhaustive grids will be filtered by the intersection between
these and \code{points_in}. Thus, the number of generated grids may be
smaller than \code{ncutsx * ncutsy}.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(nc_path)
nc_rp <- terra::spatSample(nc, 1000)
nc_gr <- par_make_grid(nc_rp, 10L, 6L)

plot(nc_rp)
plot(nc_gr, add = TRUE)
}
\seealso{
Other Parallelization: 
\code{\link{par_cut_coords}()},
\code{\link{par_fallback}()},
\code{\link{par_grid}()},
\code{\link{par_group_grid}()},
\code{\link{par_hierarchy}()},
\code{\link{par_make_gridset}()},
\code{\link{par_merge_grid}()},
\code{\link{par_multirasters}()}
}
\author{
Insang Song
}
\concept{Parallelization}
