% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_process.R
\name{par_map_args}
\alias{par_map_args}
\title{Map arguments to the desired names}
\usage{
par_map_args(fun, name_match = list(), ...)
}
\arguments{
\item{fun}{The target function to be called.}

\item{name_match}{A named list of arguments to be mapped.}

\item{...}{Arguments to be passed to the target function.}
}
\value{
The result of the target function.
}
\description{
This function maps the arguments of a target function
to the desired names. Users will use a named list \code{name_match} to
standardize the argument names, at least x and y, to the target function.
This function is particularly useful to parallelize functions for spatial
data outside \code{sf} and \code{terra} packages that do not have arguments
named x and/or y. \verb{par_*} functions could detect such functions by
wrapping nonstandardized functions to parallelize the computation.
}
\examples{
# Example target function
example_fun <- function(x, y, z = 1) {
  return(c(x = x, y = y, z = z))
}

# Example usage of map_args_xy
result <- map_args_xy(fun = example_fun,
                      name_match = list(a = "x", b = "y"),
                      a = 10, b = 20, z = 5)
print(result)
}
\seealso{
Other Helper functions: 
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{get_clip_ext}()},
\code{\link{par_def_q}()},
\code{\link{reproject_std}()},
\code{\link{reproject_to_raster}()},
\code{\link{vect_validate}()}
}
\concept{Helper functions}
