% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_grid.R
\docType{data}
\name{prediction_grid}
\alias{prediction_grid}
\title{Regular grid points in the mainland United States at 1km spatial resolution}
\format{
A data frame with 8,092,995 rows and three variables:
\describe{
\item{site_id}{Unique point identifier. Arbitrarily generated.}
\item{lon}{Longitude}
\item{lat}{Latitude}
}
}
\source{
Mainland United States polygon was obtained from
the US Census Bureau.
}
\usage{
prediction_grid
}
\description{
Regular grid points in the mainland United States at 1km spatial resolution
}
\note{
Coordinates are in EPSG:5070 (Conus Albers Equal Area)
}
\examples{
data("prediction_grid", package = "chopin")
}
\seealso{
Other Dataset: 
\code{\link{ncpoints}}
}
\concept{Dataset}
\keyword{datasets}
