% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rast_short}
\alias{rast_short}
\title{Quick call for SpatRaster with a window}
\usage{
rast_short(rasterpath, win)
}
\arguments{
\item{rasterpath}{character(1). Path to the raster file.}

\item{win}{Named integer vector (4) or terra::ext() results.}
}
\value{
SpatRaster object.
}
\description{
Quick call for SpatRaster with a window
}
\author{
Insang Song
}
