% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{rast_short}
\alias{rast_short}
\title{Quick call for SpatRaster with a window}
\usage{
rast_short(rasterpath = NULL, win = NULL)
}
\arguments{
\item{rasterpath}{character(1). Path to the raster file.}

\item{win}{Named integer vector (4) or terra::ext() results.}
}
\value{
SpatRaster object.
}
\description{
Quick call for SpatRaster with a window
}
\examples{
library(terra)
bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
ext_small <- terra::ext(
  c(xmin = -80, xmax = -76, ymin = 35, ymax = 36)
)
rast_short(bcsd_path, ext_small)
}
\author{
Insang Song
}