% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{reproject_b2r}
\alias{reproject_b2r}
\title{Align vector CRS to raster's}
\usage{
reproject_b2r(vector, raster)
}
\arguments{
\item{vector}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object}

\item{raster}{\code{SpatRaster} object}
}
\value{
Reprojected object in the same class as \code{vector}
}
\description{
Align vector CRS to raster's
}
\examples{
# NOT TO RUN
}
\seealso{
Other Helper functions: 
\code{\link{any_class_args}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext2poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{is_within_ref}()},
\code{\link{rast_short}()},
\code{\link{reproject_std}()},
\code{\link{vect_valid_repair}()}
}
\author{
Insang Song
}
\concept{Helper functions}