% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{reproject_b2r}
\alias{reproject_b2r}
\title{Reproject vectors to raster's CRS}
\usage{
reproject_b2r(vector, raster)
}
\arguments{
\item{vector}{sf/stars/SpatVector/SpatRaster object}

\item{raster}{SpatRaster object}
}
\value{
Reprojected object in the same class as \code{vector}
}
\description{
Reproject vectors to raster's CRS
}
\author{
Insang Song
}