% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{reproject_std}
\alias{reproject_std}
\title{Check coordinate system then reproject}
\usage{
reproject_std(input, crs_standard = "EPSG:4326")
}
\arguments{
\item{input}{Input object one of sf or terra::Spat* object}

\item{crs_standard}{character(1). A standard definition of
coordinate reference system. Default is \code{"EPSG:4326"}
Consult \href{https://epsg.io}{epsg.io} for details of other CRS.}
}
\value{
A (reprojected) \code{sf} or \code{SpatVector} object.
}
\description{
The input is checked whether its coordinate system is
present. If not, it is reprojected to the CRS specified in
\code{crs_standard}.
}
\note{
This function works well with EPSG codes.
}
\examples{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

base_crs <- "EPSG:5070"
nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc_sf <- sf::st_read(nc_path)
reproject_std(nc_sf, base_crs)

nc_vect <- terra::vect(nc_sf)
reproject_std(nc_vect, base_crs)
}
\author{
Insang Song
}