% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{reproject_to_raster}
\alias{reproject_to_raster}
\title{Align vector CRS to raster's}
\usage{
reproject_to_raster(vector = NULL, raster = NULL)
}
\arguments{
\item{vector}{\code{sf}/\code{stars}/\code{SpatVector}/\code{SpatRaster} object}

\item{raster}{\code{SpatRaster} object}
}
\value{
Reprojected object in the same class as \code{vector}
}
\description{
Align vector CRS to raster's
}
\examples{
library(terra)
library(sf)
options(sf_use_s2 = FALSE)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
elev <- system.file("ex/elev.tif", package = "terra")
nc <- terra::vect(ncpath)
elev <- terra::rast(elev)
reproject_to_raster(nc, elev)
}
\seealso{
Other Helper functions: 
\code{\link{check_dist_incorrect}()},
\code{\link{clip_ras_ext}()},
\code{\link{clip_vec_ext}()},
\code{\link{crs_check}()},
\code{\link{datamod}()},
\code{\link{dep_check}()},
\code{\link{dep_switch}()},
\code{\link{ext_to_poly}()},
\code{\link{get_clip_ext}()},
\code{\link{is_bbox_within_reference}()},
\code{\link{par_def_q}()},
\code{\link{par_map_args}()},
\code{\link{reproject_std}()},
\code{\link{vect_validate}()}
}
\author{
Insang Song
}
\concept{Helper functions}
\keyword{internal}
\keyword{soft-deprecated}
