% Generated by roxygen2: do not edit by hand
% Please edit documentation in scomps_rmarkdown_litr.rmd.
\name{set_clip_extent}
\alias{set_clip_extent}
\title{Setting the clipping extent}
\usage{
set_clip_extent(pnts, buffer_r)
}
\arguments{
\item{pnts}{One of sf or vect class. Target points of computation.}

\item{buffer_r}{numeric(1). Buffer radius. It is assumed in metres}
}
\value{
A terra::ext or sfc_POLYGON object of the computation extent.
}
\description{
Return clipping extent with buffer radius. It assumes the input CRS is projected and linear unit is meters.
}
\author{
Insang Song
}