% Generated by roxygen2: do not edit by hand
% Please edit documentation in chopin_rmarkdown_litr.rmd.
\name{summarize_aw}
\alias{summarize_aw}
\title{Area weighted summary using two polygon sf or SpatVector objects}
\usage{
summarize_aw(poly_in = NULL, poly_weight = NULL, id_poly_in = "ID")
}
\arguments{
\item{poly_in}{A sf/SpatVector object at weighted means will be calculated.}

\item{poly_weight}{A sf/SpatVector object from
which weighted means will be calculated.}

\item{id_poly_in}{character(1).
The unique identifier of each polygon in \code{poly_in}}
}
\value{
A data.frame with all numeric fields of area-weighted means.
}
\description{
When \code{poly_in} and \code{poly_weight} are different classes,
\code{poly_weight} will be converted to the class of \code{poly_in}.
}
\examples{
# package
library(sf)
sf_use_s2(FALSE)
# run
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
nc <- sf::st_transform(nc, 5070)
pp <- sf::st_sample(nc, size = 300)
pp <- sf::st_as_sf(pp)
pp[["id"]] <- seq(1, nrow(pp))
sf::st_crs(pp) <- "EPSG:5070"
ppb <- sf::st_buffer(pp, nQuadSegs=180, dist = units::set_units(20, 'km'))

system.time({ppb_nc_aw <- summarize_aw(ppb, nc, 'id')})
summary(ppb_nc_aw)
#### Example of summarize_aw ends ####
}
\seealso{
Other Macros for calculation: 
\code{\link{extract_at_buffer}()},
\code{\link{extract_at_poly}()},
\code{\link{extract_at}()},
\code{\link{kernelfunction}()},
\code{\link{summarize_sedc}()}
}
\author{
Insang Song \email{geoissong@gmail.com}
}
\concept{Macros for calculation}